SIstem informasi akademik
The Academic Information System (E-Learning) aimed at SMA Negeri 1 Bondowoso provides a forum for interaction between teachers and students where teachers can upload assignments or
materials to be given. While students can upload assignments that have been done. And in it there are also other features in the form of attendance, events, academic calendars, and others so that it
can make it easier for teachers or students to carry out the teaching and learning process without the need to use many platforms.